requirements.txt
absl-py==0.7.1
astor==0.8.0
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.8.0
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.1
numpy==1.16.0
opencv-python==4.1.0.25
Pillow==6.1.0
protobuf==3.9.0
pygame==1.9.6
pyparsing==2.4.2
python-dateutil==2.8.0
PyYAML==5.1.2
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
sklearn==0.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0rc1
termcolor==1.1.0
Werkzeug==0.15.5
wrapt==1.11.1
